Final version of stack-safe FoldLeft
